Clean Taxa from GBIF

Clean the taxonomic list using GBIF

Usage

Clean_Taxa_rgbif(Cleaned_Taxize, WriteFile = F, Species_Only = T)

Arguments

Cleaned_Taxize

a data frame containing the cleaned taxonomic list from function Clean_Taxa_Taxize

WriteFile

logical if False (default) only returns a data frame, if TRUE will generate a folder (Results in the Working directory with a csv of the results)

Species_Only

logical, if TRUE (default) only species will be returned, if FALSE, it will return the highest possible taxonomic resolution

Value

A csv file containing the cleaned taxonomic list

References

Chamberlain S, Barve V, Mcglinn D, Oldoni D, Desmet P, Geffert L, Ram K (2023). rgbif: Interface to the Global Biodiversity Information Facility API_ R package version 3.7.4,

Examples

Cleaned_Taxize <- Clean_Taxa_Taxize(Taxons = c("Canis lupus", "C. lupus"))
#> Joining with `by = join_by(Taxa)`
Clean_Taxa_rgbif(Cleaned_Taxize)
#> Joining with `by = join_by(matched_name2)`
#> # A tibble: 1 × 12
#> # Groups:   species [1]
#>   Taxa  matched_name2 confidence canonicalName kingdom phylum class order family
#>   <chr> <chr>              <int> <chr>         <chr>   <chr>  <chr> <chr> <chr> 
#> 1 Cani… Canis lupus           99 Canis lupus   Animal… Chord… Mamm… Carn… Canid…
#> # ℹ 3 more variables: genus <chr>, species <chr>, rank <chr>