Generate and Transform a Land-Use Suitability Table

This function takes the path to a raster file containing land-use suitability values, converts it to a data frame, and filters it to include only the cells that have a value of 1 in at least one of the land-use categories. It then transforms the filtered data into a long-format table where each row corresponds to a specific cell and habitat type.

Usage

generate_long_landuse_table(path)

Arguments

path

A string representing the file path to the raster file containing land-use suitability data.

Value

A data frame in long format with the raster cells that have a suitability value of 1 in at least one of the land-use categories. The data frame includes the cell identifier and the corresponding habitat types (Habitat) where suitability is 1.

Examples

# Get path for habitat suitability raster
HabSut <- system.file("ex/HabSut.tif", package = "SpeciesPoolR")

# Use the function to generate and transform the land-use suitability table
generate_long_landuse_table(path = HabSut)
#> Error in melt.data.table(as.data.table(DF), id.vars = "cell", measure.vars = c("ForestDryPoor",     "ForestDryRich", "ForestWetPoor", "ForestWetRich", "OpenDryPoor",     "OpenDryRich", "OpenWetPoor", "OpenWetRich"), variable.name = "Habitat",     value.name = "Suitability", na.rm = TRUE): One or more values in 'measure.vars' is invalid.